# e2e: faster ssh/scp by multiplexing socket