Benchmark results:

```
time distro=debian-sid code="exit 0" ./run_test.sh test
original: 5.8 s
fast ssh: 3.8 s

time distro=debian-sid code="vm-install-cri-resmgr-webhook" ./run_test.sh test
original: 23 s
fast ssh: 15 s
```
(The latter with PR #465 .)